+++ frontend/app/components/templates/HomePage/HomePage.tsx
@@ -9,6 +9,12 @@ import { CardContent } from '../../modules/CardContent/CardContent'
 import { Button } from '../../elements/Button/Button'
 import Link from 'next/link'
 import { CardImage } from '../../modules/CardImage/CardImage'
+import { Navigation } from 'swiper'
+import { useMediaQuery } from '@/hooks/useMediaQuery'
+/* eslint-disable import/no-unresolved */
+import { Swiper, SwiperSlide } from 'swiper/react'
+import 'swiper/css'
+import 'swiper/css/navigation'
 
 type Props = {
   analytics: Service[]
@@ -56,6 +62,9 @@ export const HomePage = ({ analytics, solutions, innovations }: Props) => {
       href: '/#new-innovations'
     }
   ]
+
+  const isMobile = useMediaQuery('(max-width: 480px)')
+
   return (
     <div className={styles.container}>
       <section className={styles.heroSection}>
@@ -76,7 +85,7 @@ export const HomePage = ({ analytics, solutions, innovations }: Props) => {
                   <h3 className={styles.secondTitle}>{item.secondTitle}</h3>
                   <p className={styles.desc}>{item.description}</p>
                   <div className={styles.arrow}>
-                    <Link href={item.href}>
+                    <Link href={item.href} passHref>
                       <img
                         className={styles.arrowImg}
                         src="/assets/icons/arrow-down.svg"
@@ -128,17 +137,38 @@ export const HomePage = ({ analytics, solutions, innovations }: Props) => {
         id="solutions"
         className={`${styles.solutionsSection} ${styles.sectionPadding}`}>
         <h1 className={styles.titleLine}>Our Solutions</h1>
-        <div className={styles.cards}>
+        <Swiper
+          modules={[Navigation]}
+          className={styles.swipper}
+          navigation={!isMobile}
+          centeredSlides={isMobile}
+          slidesPerView={'auto'}
+          spaceBetween={10}
+          grabCursor
+          breakpoints={{
+            '480': {
+              slidesPerView: 2,
+              spaceBetween: 20
+            },
+            '768': {
+              slidesPerView: 3,
+              spaceBetween: 20
+            }
+          }}>
           {solutions.map((solution) => (
-            <CardFull
-              key={solution.id}
-              img={`/assets/images/solutions/${solution.thumbnail}`}
-              title={solution.name}
-              desc={solution.short_description}
-              href={`/solutions/${solution.slug}`}
-            />
+            <SwiperSlide className={styles.swipperSlide} key={solution.id}>
+              <CardFull
+                img={`/assets/images/solutions/${solution.thumbnail}`}
+                title={solution.name}
+                href={`/solutions/${solution.slug}`}>
+                <div className={styles.content}>
+                  <h3>{solution.name}</h3>
+                  <p>{solution.short_description}</p>
+                </div>
+              </CardFull>
+            </SwiperSlide>
           ))}
-        </div>
+        </Swiper>
       </section>
       <section className={styles.intersection}>
         <div className={styles.container}>